requirements.txt
bs4
langchain_mistralai
langchain
langchain_community
python_dotenv
duckduckgo-search
sentence-transformers
faiss-cpu